+++ sender/src/main/java/dmytro/kudriavtsev/sender/dtos/ExchangeDTO.java
@@ -1,12 +1,14 @@
 package dmytro.kudriavtsev.sender.dtos;
 
+import java.math.BigDecimal;
+
 public class ExchangeDTO {
     private String email;
     private String event;
     private String firstCurrency;
-    private double firstSum;
+    private BigDecimal firstSum;
     private String secondCurrency;
-    private double secondSum;
+    private BigDecimal secondSum;
     private boolean success;
 
     public ExchangeDTO() {
@@ -36,11 +38,11 @@ public class ExchangeDTO {
         this.firstCurrency = firstCurrency;
     }
 
-    public double getFirstSum() {
+    public BigDecimal getFirstSum() {
         return firstSum;
     }
 
-    public void setFirstSum(double firstSum) {
+    public void setFirstSum(BigDecimal firstSum) {
         this.firstSum = firstSum;
     }
 
@@ -52,11 +54,11 @@ public class ExchangeDTO {
         this.secondCurrency = secondCurrency;
     }
 
-    public double getSecondSum() {
+    public BigDecimal getSecondSum() {
         return secondSum;
     }
 
-    public void setSecondSum(double secondSum) {
+    public void setSecondSum(BigDecimal secondSum) {
         this.secondSum = secondSum;
     }
